+++ source/features/open-all-notifications.tsx
@@ -11,7 +11,7 @@ import features from '../feature-manager.js';
 import {getIdentifiers} from '../helpers/feature-helpers.js';
 import openTabs from '../helpers/open-tabs.js';
 import observe from '../helpers/selector-observer.js';
-import {tooltipped} from '../github-helpers/tooltip.js';
+import {tooltipped} from '../helpers/tooltip.js';
 import {appendBefore} from '../helpers/dom-utils.js';
 
 // Selector works on:
