+++ source/features/pr-commit-lines-changed.tsx
@@ -6,7 +6,7 @@ import features from '../feature-manager.js';
 import api from '../github-helpers/api.js';
 import pluralize from '../helpers/pluralize.js';
 import observe from '../helpers/selector-observer.js';
-import {tooltipped} from '../github-helpers/tooltip.js';
+import {tooltipped} from '../helpers/tooltip.js';
 import GetCommitChanges from './pr-commit-lines-changed.gql';
 
 // Adapted from GitHub https://github.com/refined-github/refined-github/pull/9486#discussion_r3252807259
