+++ source/features/table-input.tsx
@@ -11,7 +11,7 @@ import features from '../feature-manager.js';
 import {actionBar} from '../github-helpers/selectors.js';
 import observe from '../helpers/selector-observer.js';
 import smartBlockWrap from '../helpers/smart-block-wrap.js';
-import {withTooltipRef} from '../github-helpers/tooltip.js';
+import {withTooltipRef} from '../helpers/tooltip.js';
 
 function addTable({delegateTarget: square}: DelegateEvent<MouseEvent, HTMLButtonElement>): void {
 	const container = closestElementOptional('fieldset', square) // Issue
